+++ package.json
@@ -20,6 +20,7 @@
     "feather-icons": "^4.29.0",
     "react": "18.2.0",
     "react-native": "^0.70.6",
+    "react-native-axios": "^0.17.1",
     "react-native-draggable": "^3.3.0",
     "react-native-image-lazy-loading": "^1.0.4",
     "react-native-incall-manager": "^4.0.1",
