+++ src/ReactQuery/getCollectionItems.js
@@ -1,0 +1,24 @@
+
+//   const _getCollectionItems = () => {
+//     try {
+//       fetch(
+//         `https://api.sendwish.link:8081/collection/${nickName}/${collectionId}`,
+//         {
+//           method: 'GET',
+//           headers: {'Content-Type': 'application/json'},
+//         },
+//       )
+//         .then(res => {
+//           return res.json();
+//         })
+//         .then(data => {
+//           console.log('쿼리 컬렉션 아이템', data);
+//         });
+//     } catch (e) {
+//       return e;
+//     }
+//   };
+
+
+// export default _getCollectionItems;
+
