+++ src/screens/SignIn.js
@@ -85,7 +85,7 @@ const SignIn = ({navigation}) => {
       const value = await AsyncStorage.getItem('userdata')
 
       if(value !== null){
-        console.log('storage date check userdata!!!', value);
+        // console.log('storage date check userdata!!!', value);
         navigation.navigate('Navigation',{screen:'Main', params:{nickName: JSON.parse(value).nickName, accessToken: JSON.parse(value).accessToken}})
       }
     }
@@ -96,15 +96,15 @@ const SignIn = ({navigation}) => {
   getData();
 
   // 토큰이 없으면 로그인 화면으로 가는지 확인용으로 storage clear test
-  const clearData = async () => {
-    try{
-      await AsyncStorage.clear();
-    }
-    catch(e){
-      console.log(e);
-    }
-  }
-  clearData();
+  // const clearData = async () => {
+  //   try{
+  //     await AsyncStorage.clear();
+  //   }
+  //   catch(e){
+  //     console.log(e);
+  //   }
+  // }
+  // clearData();
 
   useEffect(() => {
     setDisabled(!(nickName && password && !errorMessage));
